+++ modules/core/src/utils.ts
@@ -1,25 +1,5 @@
-import {getSerializable, SerializableProperty} from "./decorator/serialize.decorator";
-
-if (typeof Object.assign !== "function") {
-    Object.assign = function (target) {
-        "use strict";
-        if (target === null) {
-            throw new TypeError("Cannot convert undefined or null to object");
-        }
-        target = Object(target);
-        for (var index = 1; index < arguments.length; index++) {
-            var source = arguments[index];
-            if (source !== null) {
-                for (var key in source) {
-                    if (Object.prototype.hasOwnProperty.call(source, key)) {
-                        target[key] = source[key];
-                    }
-                }
-            }
-        }
-        return target;
-    };
-}
+import {AsyncValidatorFn, Validators, ValidatorFn} from "@angular/forms";
+import {getSerializables, SerializableProperty} from "./decorator/serialize.decorator";
 
 export function isDefined(object: any): boolean {
     return object !== undefined && object !== null;
@@ -31,6 +11,18 @@ export function isEmptyString(_string: string | null | undefined): boolean {
 
 export function isFunction(object: any): boolean {
     return typeof object === "function";
+}
+
+export function getFunctionName(func: any): string | null {
+
+    if (func.name) {
+        return func.name;
+
+    } else {
+
+        let name = /^function\s+([\w\$]+)\s*\(/.exec(func.toString());
+        return name ? name[1] : null;
+    }
 }
 
 export function getValue(object: any, key: string, defaultValue: any): any {
@@ -61,10 +53,62 @@ export function getValue(object: any, key: string, defaultValue: any): any {
     return value;
 }
 
-export function serialize(context): Object {
 
-    return getSerializable(context).reduce((prev, prop: SerializableProperty) => {
-        prev[prop.name] = context[prop.key];
+export function serializeValidator(validator: ValidatorFn | AsyncValidatorFn): string {
+
+    for (let validatorName in Validators) {
+
+        if (Validators.hasOwnProperty(validatorName) && validator === Validators[validatorName]) {
+            return validatorName;
+        }
+    }
+
+    return null;
+}
+
+export function serializeValidators(validators: Array<ValidatorFn | AsyncValidatorFn>): Array<string> {
+
+    let serialized = [];
+
+    validators.forEach(validator => {
+
+        let validatorName = serializeValidator(validator);
+
+        if (validatorName) {
+            serialized.push(validatorName);
+        }
+    });
+
+    return serialized;
+}
+
+export function deserializeValidator(serialized: string): ValidatorFn | AsyncValidatorFn {
+
+    return Validators[serialized];
+}
+
+export function deserializeValidators(serialized: Array<string>): Array<ValidatorFn | AsyncValidatorFn> {
+
+    return serialized.map(validatorName => deserializeValidator[validatorName]);
+}
+
+export function serialize(target, prototype?): Object {
+
+    return getSerializables(prototype || target).reduce((prev, prop: SerializableProperty) => {
+
+        if (prop.key === "validators" || prop.key === "asyncValidators") {
+
+            prev[prop.name] = serializeValidators(target[prop.key]);
+
+        } else if (prop.key === "validator" || prop.key === "asyncValidator") {
+
+            prev[prop.name] = serializeValidator(target[prop.key]);
+
+        } else {
+            prev[prop.name] = target[prop.key];
+        }
+
         return prev;
+
     }, {});
 }